+++ StswExpress.Commons/Utils/Collections/_items/_interfaces/IStswTrackableItem.cs
@@ -2,7 +2,7 @@
 
 namespace StswExpress.Commons;
 /// <summary>
-/// Defines a contract for items that can be tracked for changes in state and property notifications.
+/// Defines a contract for items that can be tracked for changes in state.
 /// </summary>
 public interface IStswTrackableItem : INotifyPropertyChanged
 {
@@ -10,9 +10,4 @@ public interface IStswTrackableItem : INotifyPropertyChanged
     /// Gets or sets the state of the trackable item.
     /// </summary>
     public StswItemState ItemState { get; set; }
-
-    /// <summary>
-    /// Gets or sets a value indicating whether to show details for the item.
-    /// </summary>
-    public bool? ShowDetails { get; set; }
 }